% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp_tnbc_A2C9}
\alias{exp_tnbc_A2C9}
\title{RNA-Seq of a TNBC patient}
\format{A data frame with 20501 rows and 2 columns:
\describe{
  \item{Tumor_Sample}{RNA-Seq data of a tumor sample}
  \item{Healthy_Sample}{RNA-Seq data of a healthy sample}
  ...
}}
\source{
\url{https://portal.gdc.cancer.gov/}
}
\usage{
exp_tnbc_A2C9
}
\description{
A dataset containing RNA-Seq data of two transcriptomes belonging to a triple negative breast
cancer (TNBC) patient
}
\keyword{datasets}
