% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iDEG.R
\name{iDEG}
\alias{iDEG}
\title{Identification of \emph{i}ndividualized \emph{d}ifferentially
\emph{e}xpressed \emph{g}enes (iDEG).}
\usage{
iDEG(baseline, case, cutOff = 0.2, normalization = F,
  dataDistribution = c("NB", "Poisson"), numBin = 100, estSize = F,
  rankBaseline = T, estBaseline = F, constDisp = T, spar = NULL,
  df = 7, nulltype = 1, pct = 1e-04, plot = 0)
}
\arguments{
\item{baseline}{A vector of gene expression levels of the baseline
transcriptome (e.g., healthy tissue)}

\item{case}{A vector of gene expression levels of the case transcriptome (e.g., tumor tissue)}

\item{normalization}{A logical variable indicating if normalization has been done}

\item{numBin}{Number of bins used to group all genes into. Default is 100.}

\item{rankBaseline}{If True, iDEG groups all genes based on the gene expression levels of the baseline transcriptome. If False, iDEG group all genes based on the gene expression levels of the average of baseline and case transcriptomes.}

\item{constDisp}{If True, iDEG assumes the dispersion is a count across all genes. If False, iDEG assume dispersion is a smooth fucntion os expression mean}

\item{df}{the degrees of freedom used for estimating marginal distrution.}

\item{nulltype}{Type of null distribution assumed in computing the probability of gene differential expression. 0 is the theoretical null N(0,1), 1 is maximum likelihood estimation.}

\item{pct}{the percentage of genes exculded from fiting the two-group mixture model.}

\item{dataDistribuiton}{The distribuitonal assumption of the RNA-Seq data under analysis. Possible values are 'Poisson' and 'NB'. Default is NB--negative binomial.}
}
\value{
'iDEG' produces a list containing the following elements:

\describe{
  \item{results}{a table iDEG result for each gene. The first two columns are
the gene epxression values of the two transcriptomes provided by the user.
The thrid column is the local false discovery rate, which provides the
probability of a gene being differentially expresseed. The fourth column is
the statistic used to compute the local false discovery rate, and can be used
as an effect size.}
\item{sizeHat}{When the assumptioin of constant dispersion across genes is
made, this is an single estimate of the common dispersion. When the
assumptioin of non-constant is made, this is a vector of estimates for the
dispersion parameter of each gene.}
}
}
\description{
Identify differentionally expressed genes between two conditions,
and only one transcriptome is collected for each condition.
}
\examples{
set.seed(1)
exp_mean1 <- rexp(20000, 1/500) + 1
exp_mean2 <- exp_mean1
exp_mean2[1:100] <- exp_mean2[1:100] * 10
transcriptome1 <- rnbinom(n = length(exp_mean1), size = 60, mu = exp_mean1)
transcriptome2 <- rnbinom(n = length(exp_mean2), size = 60, mu = exp_mean2)
res <- iDEG(transcriptome1,transcriptome2)

}
